refactor: shrink public API surface for v3
