+++ docs/02_concepts/code/10_architecture_imports.py
@@ -1,4 +1,4 @@
-from apify_client import (
+from apify_client.http_clients import (
     HttpClient,
     HttpClientAsync,
     HttpResponse,
