+++ docs/03_guides/code/05_custom_http_client_sync.py
@@ -1,10 +1,14 @@
 from __future__ import annotations
 
-from typing import Any
+from typing import TYPE_CHECKING, Any
 
 import httpx
 
-from apify_client import ApifyClient, HttpClient, HttpResponse, Timeout
+from apify_client import ApifyClient
+from apify_client.http_clients import HttpClient, HttpResponse
+
+if TYPE_CHECKING:
+    from apify_client.types import Timeout
 
 TOKEN = 'MY-APIFY-TOKEN'
