+++ src/apify_client/__init__.py
@@ -1,25 +1,11 @@
 from importlib import metadata
 
 from ._apify_client import ApifyClient, ApifyClientAsync
-from ._http_clients import (
-    HttpClient,
-    HttpClientAsync,
-    HttpResponse,
-    ImpitHttpClient,
-    ImpitHttpClientAsync,
-)
-from ._types import Timeout
 
 __version__ = metadata.version('apify-client')
 
 __all__ = [
     'ApifyClient',
     'ApifyClientAsync',
-    'HttpClient',
-    'HttpClientAsync',
-    'HttpResponse',
-    'ImpitHttpClient',
-    'ImpitHttpClientAsync',
-    'Timeout',
     '__version__',
 ]
