+++ src/apify_client/_resource_clients/build_collection.py
@@ -11,7 +11,7 @@ if TYPE_CHECKING:
     from collections.abc import AsyncIterator, Iterator
 
     from apify_client._models import BuildShort
-    from apify_client._types import Timeout
+    from apify_client.types import Timeout
 
 
 @docs_group('Resource clients')
