+++ src/apify_client/_resource_clients/log.py
@@ -11,8 +11,8 @@ from apify_client.errors import ApifyApiError
 if TYPE_CHECKING:
     from collections.abc import AsyncIterator, Iterator
 
-    from apify_client._http_clients import HttpResponse
-    from apify_client._types import Timeout
+    from apify_client.http_clients import HttpResponse
+    from apify_client.types import Timeout
 
 
 @docs_group('Resource clients')
