+++ tests/unit/test_pluggable_http_client.py
@@ -6,22 +6,22 @@ from typing import TYPE_CHECKING, Any
 import pytest
 
 import apify_client as apify_client_module
-from apify_client import (
-    ApifyClient,
-    ApifyClientAsync,
+from apify_client import ApifyClient, ApifyClientAsync
+from apify_client.errors import ApifyApiError
+from apify_client.http_clients import (
     HttpClient,
     HttpClientAsync,
     HttpResponse,
+    ImpitHttpClient,
+    ImpitHttpClientAsync,
 )
-from apify_client._http_clients import ImpitHttpClient, ImpitHttpClientAsync
-from apify_client.errors import ApifyApiError
 
 if TYPE_CHECKING:
     from collections.abc import AsyncIterator, Iterator
 
     from pytest_httpserver import HTTPServer
 
-    from apify_client._types import Timeout
+    from apify_client.types import Timeout
 
 
 # -- Test response and client implementations --
@@ -287,12 +287,11 @@ async def test_apify_client_async_with_custom_http_client_accepts_url_params() -
 
 
 def test_public_exports() -> None:
-    """Test that all HTTP client types are available from the public API."""
-    assert hasattr(apify_client_module, 'HttpClient')
-    assert hasattr(apify_client_module, 'HttpClientAsync')
-    assert hasattr(apify_client_module, 'HttpResponse')
-    assert hasattr(apify_client_module, 'ImpitHttpClient')
-    assert hasattr(apify_client_module, 'ImpitHttpClientAsync')
+    """HTTP client types are exposed from `apify_client.http_clients`, not the root namespace."""
+    http_clients_module = apify_client_module.http_clients
+    for name in ('HttpClient', 'HttpClientAsync', 'HttpResponse', 'ImpitHttpClient', 'ImpitHttpClientAsync'):
+        assert hasattr(http_clients_module, name)
+        assert not hasattr(apify_client_module, name)
 
 
 # -- http_client property --
